refactor: split harness avatar and council modules
